refactor remove skill space to be shared
